feat: created commons package, tried to send messages from wearOs to Android.
A commons package was created to have shared or common code to both
android and wear os. So many commons files were imported and used in
different projects.

The last thing I did is try to send a message from WearOs to android.
It seems like WearOs is sending the message correctly, but Android
is not receiving this message. We need to debug much more.