feat: resolved communication between android activity and android service
It was created in DataTransmissionService a method that allows us
to send messages via Broadcast. We recive this messages implementing
a LocalBroadcast in the MainActivity. The SensorDataMessage was updated,
now SensoreMessage implements Parcelabe and we use it as a unit to share
messages from WearOS to Android Service and from Android Service to
Android Activity