feat: completed locations options activity
Now we can from the location options activity see the known
or registered locations. We can see them in two main formats,
as a list or in a map. We have an available float button that
drives us to the add location activity. In this activity we have
a form and we can fill the gaps and mark a position in the available
map to add a new known location.