feat: Added an Adl detection service that comunicates bidirectionally
We've just created a new Service that extends Service class. This will
be the responsable of receiving data sensors and process them searching
ADLs. The most important thing we got from this new service is that we've
tested that is possible to send messages from a service to an activity
and from an activity to a service using the same methods (registering a
receiver in the LocalBroadCaster).