feat: created new ADL service and estblished multi data communication
Now we can send an array of data avoiding messages overload. The
ADL detection service receives all this data and is able to process
it in an easy way. Next step will be to get together all data
that was taken in a close time period (for example in the same second).