feat: implemented data filter and some rules
One Rule was implemented and it's necessary to fix some aspects, we
have to understand the ratio of reading of each sensor. Sensors like
accelerometer read data each millisecond and we are then sending a
buffer full of only one data (the last acceleromter). That's not well.
We have to fix this.