feat: added bluetooth devices provider
This new class scans the environmente looking for
visible and close devices using bluetooth. It also
display this devices. It's necessary to create a
background task or something or at least try to use
a different function to have this always working on background.

Maybe a good idea it would be to use the start discovery
option.