feat: created local database structure and added BluetoothDevices entity
we've created a new database structure in the project using
it's own package. We added a new table (entity) for detected bluetooth
devices and we use it correctly. We also tested if it's possible to use
this database from a service. It's not possible or at least easy to
implement, in fact the database room is a service itself.