feat: removed local database service after research
We tried to create a new service but it wasn' possible.
We're realizing that some times it doesn't make sense to
create a service, for example when a service uses the
context of the app or hardware resources or calls to
an API that uses an already implemented service, we should
then use the existing service or the context app, etc.