feat: reorganized structure of project
We've created a new package for activitities and another one
for services so now we can see more clear what we want to use
or edit. We've also created an util packages that contains
a lot of classes we created as data structures to use in other
activities or services or both.