feat: created a new entity for detected adl sensors
A new entity was created and integrated in the activity
were it was necessary. Now we are loading content from
database using this entity, next step is write content
on it but we have a service to do this so we need to
use this new entity and its DAO from the mentioned service.

refs: #TFG-61