feat: added a new entity for detected adls and integrted changes
A new table was added to our local mobile database and
now we use it where it was necessary. We integrated the
new entity usage in the adl detection service. We also
discovered that we can use hardware resources always while
having permissions and not having to ask for them.

refs: #TFG-58