feat: redesigned bluetooth detection service
Now this service uses another kind of services
and just awaits to receive some events related to
found devices. Theremore, the collected detecttion devices
data is stored in the local database and the adl detection
service can use those data to know if the adls are being
detected accompanied or not.

refs: #TFG-17