feat: added java mqtt client and a database service
Currently we've looked for a good api to implement
a mqtt client for android studio, the best option is hivemq
mqtt client. We tested it and it worked fine, but there are
some errors to fix. In the server side the database service
was created but it doesn't do nothing interesting yet. Proximately
it will connect to a database and receive information from the subscribers.

refs: #TFG-33