feat: completed implementation of backup service
We call backup service to a service that is going to be
awaiting a certain amount of time to fire an event and backup
sensors data stored in our local database. We've tested schedulers
in android and we've tested AlarmManager, the final result wasn't
what we expected so we've decided to use a Handle with post delay.
This is a way to execute repeatedly a certain task exactly when we want,
i.e. each 5 minutes.

refs: #TFG-33