feat: adapted the adl detection service to the new extensible adl system
We extract the adls and the events that form them from a remote sql
database service through mqtt. Once we get them we save them in our
local database an update a copy of that new data we keep in memory.
So when a new bunch of sensors reads is received by the adl detection
service, it uses the mentioned copy in memory to evaluate this adls.
So each adl keeps its events states (true or false) and it tries to
evaluate each event associated to an adl always respecting the order.
Once all the events of an adl were evaluated then we can evaluate the
entire adl state based on its events states. If all events are true
then the adl is true too. Then we run another function to unmount the
adl state, this goes from the last event to the first and it checks if
each one of the adls in a true state stills being true or not. If any
of theese is false (respecting always the order of course) then the adl
is no longer detected.

refs: #TFG-92 #TFG-87