fix: resolved unmounting adl detection function
When we detect an adl, we have to "mount" it, this means:
to check every event associated to the and climb along them as
if it was a mountain, so we mount it. Once we have detected the
adl, then we need to unmount this mountain, this means to invalidate
the events properly. The current unmount function doesn't respect
any particular order, just checks events from start to end
and if finds any false events then invalidates all the posteriors.

refs: #TFG-92 #TFG-87