feat: moved testMqtt from mainactivity to adl detection service
The mentioned method was moved as it was said. The service for
adl detection uses it correctly and is adapted to its use. However,
it's necessary to add a mqtt publish request for the first time the
service is launched, in that case it doesn't have any content stored
in the local database and it needs it to operate, so it would be
always waiting.

refs: #TFG-94