fix: moved method of update of adls in database.
In the last version this method was just a test in the main activity.
Now it was moved to the adl detection service. It was also
fixed the local database scheme because previously it was always adding
the same data and replacing the old by the new insertion, but not
restoring the database value. So if we would like to remove some
adl and forget them it wouldn't be possible, now when the remote service
of database informs about new adls we copy this exact content into our
local database, this means we copy the remote exactly as it is.

refs: #TFG-94 #TFG-98