test: we've tested the availability of the property responseTopic.
Now we have discovered that we can send a publish message using the responseTopic
field. So when we send this field, then we are notifying the subscribers
that they have to response publishing to the sent response topic. This is ideal
to do requests from android client to remote node.js services.

refs: #TFG-97