feat: added custom adls to users, users and reorganized the project.
Now the android client can gets through mqtt two kind of adls scheme.
The generic adls are those that the non-registered users will use.
The custom adls are those that the registered users and logged in the
system are going to use, this adls contains also the generic adls.
The services platform was reorganized again, so many logic was contained
into the database-service (now is just a database-client), all this logic
was splited in services.

refs: #TFG-97