feat: created two charts in detailed sensor list activity.
One chart is for real time heart rate and the other one is for
steps history. Currenty the data received by the charts is being
generated in real time from the sensors provided by the differnt
devices. In the case of the step counter chart, the data is hardcoded
and it's just the front end what was designed.

refs: #TFG-106