feat: splitting functionality of main activity between new services
It's not yet finishe, we'll have a services manager services that is going to
be the responsible of throwing all the other services and stopping them when it's necessary.
Also, we'll have a data collector service, this will be the responsible of saving data
into the local database, so we'll have a service that throws all the other services, services
that generate sensor data, a service that saves all this data, a service that processes all
this data to detect ADLs. This is a much better design than the previous.

refs: #MIOT-60