fix: services throwing moved to ManagerService
We've moved the tasks related to throw a new service from
main activity to a new manager service. We've also created
a constant class to encapsulate the intent filters needed for
the broadcast communication. Main activity is now a bit smaller
and it has less responsabilities, just to throw the manager service
and to show some data. Storage responsabilities were moved to the
backup service now called StorageService because it receives data
and stores it and also sends it to the remote database.

refs: #MIOT-60