feat: support ofline sensors backup
Now the backup service will check if the saved data
was correctly sent before removing it from the local database.
This is a safer way to collect data. We've also changed the way we
remove the data from the database because preiously we were deleting
all the collected data after senging just one part and not all the elements.

refs: #MIOT-62