test: secure recorded activities data sending
We were trying to secure the sending of this data and we aboard the
problem of an unexpected interruption when recording the activity.
We tried to detect a non finished activity but this is not possible
due to database updates: our updates do the next: remove all old data,
add all new data. We've to improve the updating method.

refs: #MIOT-62