feat: added activity registry interruption support
If we are experiment with the users and then the app closes
unexpectedly we won't lose the data. Now there's a registry
that persist over interruptions and asks the user if it's
necessary to follow experiment with the last activity or maybe
start a new one. We've also implemented a new feedback control,
before we've a problem when scrolling down and up because
when the pressed steps disappeared from the screen the
pressed feedback (gray color) was being lost. Now it persist
among scrolls.

refs: #MIOT-63 #MIOT-64