test: the new foreground service was tested
It was also tested and fix a new bug we found, this bug was related
to the sensorsProviderService. When we close the app then appear a message
saying: "sensorable has stopped". This was caused because the mentioned service
needed to receive an extra Integer as part of the initial intent indicating the
device that is providing the sensors, in the end it wasn't possible to work using
that because when the app is closed no intent is sent using the correct data and
it made the app crashing. Now is fixed.

refS: #MIOT-70